requirements.txt
django 
djangorestframework 
apscheduler 
requests 
youtube-search-python
